+++ Specs/MoPubSDKSpecs/Fakes/FakeMPInstanceProvider.m
@@ -7,7 +7,7 @@
 
 #import "FakeMPInstanceProvider.h"
 #import <MediaPlayer/MediaPlayer.h>
-#import "MPAdWebView.h"
+#import "MPWebView.h"
 #import "FakeMPTimer.h"
 #import "MRBundleManager.h"
 #import <FBAudienceNetwork/FBAudienceNetwork.h>
@@ -154,14 +154,17 @@
 }
 #pragma mark - HTML Ads
 
-- (MPAdWebView *)buildMPAdWebViewWithFrame:(CGRect)frame delegate:(id<UIWebViewDelegate>)delegate
+- (MPWebView *)buildMPWebViewWithFrame:(CGRect)frame delegate:(id<MPWebViewDelegate>)delegate
 {
-    if (self.fakeMPAdWebView) {
-        self.fakeMPAdWebView.frame = frame;
-        self.fakeMPAdWebView.delegate = delegate;
-        return self.fakeMPAdWebView;
+    if (self.fakeMPWebView) {
+        self.fakeMPWebView.frame = frame;
+        self.fakeMPWebView.delegate = delegate;
+        return self.fakeMPWebView;
     } else {
-        return [super buildMPAdWebViewWithFrame:frame delegate:delegate];
+        MPWebView *newWebView = [[MPWebView alloc] initWithFrame:frame];
+        newWebView.delegate = delegate;
+
+        return newWebView;
     }
 }
 
@@ -176,7 +179,7 @@
 
 #pragma mark - MRAID
 
-- (MPClosableView *)buildMRAIDMPClosableViewWithFrame:(CGRect)frame webView:(UIWebView *)webView delegate:(id<MPClosableViewDelegate>)delegate
+- (MPClosableView *)buildMRAIDMPClosableViewWithFrame:(CGRect)frame webView:(MPWebView *)webView delegate:(id<MPClosableViewDelegate>)delegate
 {
     if (self.fakeMRAIDMPClosableView != nil) {
         return self.fakeMRAIDMPClosableView;
@@ -213,7 +216,7 @@
                      }];
 }
 
-- (MRBridge *)buildMRBridgeWithWebView:(UIWebView *)webView delegate:(id<MRBridgeDelegate>)delegate
+- (MRBridge *)buildMRBridgeWithWebView:(MPWebView *)webView delegate:(id<MRBridgeDelegate>)delegate
 {
     return [self returnFake:self.fakeMRBridge
                      orCall:^{
@@ -221,11 +224,9 @@
                      }];
 }
 
-- (UIWebView *)buildUIWebViewWithFrame:(CGRect)frame
+- (MPWebView *)buildMPWebViewWithFrame:(CGRect)frame
 {
-    return [self returnFake:self.fakeUIWebView orCall:^id{
-        return [super buildUIWebViewWithFrame:frame];
-    }];
+    return [self buildMPWebViewWithFrame:frame delegate:nil];
 }
 
 - (MRVideoPlayerManager *)buildMRVideoPlayerManagerWithDelegate:(id<MRVideoPlayerManagerDelegate>)delegate