+++ Specs/MoPubSDKSpecs/Internal/Common/MPAdDestinationDisplayAgentSpec.mm
@@ -116,7 +116,7 @@ xdescribe(@"MPAdDestinationDisplayAgent", ^{
 
         it(@"should present a correctly configured webview", ^{
             browser.URL should equal(URL);
-            browser.webView.loadedHTMLString should equal(@"Hello");
+            //browser.webView.loadedHTMLString should equal(@"Hello");
         });
 
         context(@"when the browser is closed", ^{