+++ Specs/MoPubSDKSpecs/Internal/Common/MPURLResolverSpec.mm
@@ -122,7 +122,7 @@ describe(@"MPURLResolver", ^{
 
         describe(@"when there is a problem fetching the contents of the URL", ^{
             beforeEach(^{
-                url = [NSURL URLWithString:@"http://www.google.com/"];
+                url = [NSURL URLWithString:@"https://www.google.com/"];
             });
 
 
@@ -438,20 +438,20 @@ describe(@"MPURLResolver", ^{
 
             beforeEach(^{
                 malformedURLs = @[
-                                  [NSURL URLWithString:@"http://itunes.apple.com/lb/anything/in/the/world/can_go-here/id1138a2?mt=8&cg=2"],
-                                  [NSURL URLWithString:@"http://itunes.apple.com/lb/flubber"],
-                                  [NSURL URLWithString:@"http://itunes.apple.com/idiotic"],
-                                  [NSURL URLWithString:@"http://itunes.apple.com/us/id"],
-                                  [NSURL URLWithString:@"http://itunes.apple.com/us/id2123/not-valid-actually"],
-                                  [NSURL URLWithString:@"http://itunes.apple.com/us/album?id=132ab3&floop=132"],
-                                  [NSURL URLWithString:@"http://itunes.apple.com/us/floop?id=132ab&marb"],
-                                  [NSURL URLWithString:@"http://itunes.apple.com/us/?id=132ab&"],
-                                  [NSURL URLWithString:@"http://phobos.apple.com/us/WebObjects/MZStore.woa/wa/viewSoftware?id=132ab3&floop=132"],
-                                  [NSURL URLWithString:@"http://phobos.apple.com/us/floop?id=132ab&marb"],
-                                  [NSURL URLWithString:@"http://phobos.apple.com/"],
-                                  [NSURL URLWithString:@"http://phobos.apple.com/WebObjects"],
-                                  [NSURL URLWithString:@"http://phobos.apple.com/us/?foo=bar"],
-                                  [NSURL URLWithString:@"http://newton.apple.com/us/id2123"]
+                                  [NSURL URLWithString:@"https://itunes.apple.com/lb/anything/in/the/world/can_go-here/id1138a2?mt=8&cg=2"],
+                                  [NSURL URLWithString:@"https://itunes.apple.com/lb/flubber"],
+                                  [NSURL URLWithString:@"https://itunes.apple.com/idiotic"],
+                                  [NSURL URLWithString:@"https://itunes.apple.com/us/id"],
+                                  [NSURL URLWithString:@"https://itunes.apple.com/us/id2123/not-valid-actually"],
+                                  [NSURL URLWithString:@"https://itunes.apple.com/us/album?id=132ab3&floop=132"],
+                                  [NSURL URLWithString:@"https://itunes.apple.com/us/floop?id=132ab&marb"],
+                                  [NSURL URLWithString:@"https://itunes.apple.com/us/?id=132ab&"],
+                                  [NSURL URLWithString:@"https://phobos.apple.com/us/WebObjects/MZStore.woa/wa/viewSoftware?id=132ab3&floop=132"],
+                                  [NSURL URLWithString:@"https://phobos.apple.com/us/floop?id=132ab&marb"],
+                                  [NSURL URLWithString:@"https://phobos.apple.com/"],
+                                  [NSURL URLWithString:@"https://phobos.apple.com/WebObjects"],
+                                  [NSURL URLWithString:@"https://phobos.apple.com/us/?foo=bar"],
+                                  [NSURL URLWithString:@"https://newton.apple.com/us/id2123"]
                                   ];
             });
 
@@ -486,7 +486,7 @@ describe(@"MPURLResolver", ^{
     describe(@"when the URL is a redirect to something else", ^{
         __block NSURLConnection *lastConnection;
         beforeEach(^{
-            url = [NSURL URLWithString:@"http://i.will.redirect/"];
+            url = [NSURL URLWithString:@"https://i.will.redirect/"];
         });
 
         context(@"when the final URL should be opened in an in-app browser", ^{
@@ -494,7 +494,7 @@ describe(@"MPURLResolver", ^{
                 lastConnection = [NSURLConnection lastConnection];
                 lastConnection.request.URL should equal(url);
 
-                NSURL *redirectURL = [NSURL URLWithString:@"http://i.am.a.webpage"];
+                NSURL *redirectURL = [NSURL URLWithString:@"https://i.am.a.webpage"];
                 NSURLRequest *redirectRequest = [NSURLRequest requestWithURL:redirectURL];
                 NSURLRequest *approvedRedirectRequest = [lastConnection.delegate connection:lastConnection willSendRequest:redirectRequest redirectResponse:nil];
                 approvedRedirectRequest should equal(redirectRequest);
@@ -531,7 +531,7 @@ describe(@"MPURLResolver", ^{
 
     describe(@"when canceled", ^{
         beforeEach(^{
-            url = [NSURL URLWithString:@"http://www.google.com/"];
+            url = [NSURL URLWithString:@"https://www.google.com/"];
         });
 
         it(@"should cancel the connection", ^{